Implement a plugin system and load plugins into the store
