+++ dfu/api/plugin.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class DfuPlugin(ABC):
+    @abstractmethod
+    def handle(self):
+        pass
