+++ dfu/commands/__init__.py
@@ -4,3 +4,4 @@ from dfu.commands.create_package import create_package
 from dfu.commands.create_snapshot import create_snapshot
 from dfu.commands.diff import abort_diff, begin_diff, continue_diff
 from dfu.commands.load_config import get_config_paths, load_config
+from dfu.commands.load_store import load_store
